requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.29
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.5
typing-extensions==4.13.2
